heureum-platform/pyproject.toml
[tool.poetry]
name = "heureum-platform"
version = "0.1.0"
description = "Django proxy and message storage system"
authors = ["Heureum AI"]
readme = "README.md"
packages = [{include = "heureum_platform"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.0"
djangorestframework = "^3.15.0"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
httpx = "^0.28.0"
psycopg2-binary = "^2.9.10"
gunicorn = "^24.0.0"
django-allauth = "^65.14.1"
azure-communication-email = "^1.1.0"
pyjwt = "^2.11.0"
firebase-admin = "^7.1.0"
pywebpush = "^2.3.0"
croniter = "^6.0.0"
django-storages = {extras = ["azure"], version = "^1.14.6"}
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.9.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.0.0"
ruff = "^0.9.0"
mypy = "^1.13.0"
django-stubs = "^5.1.3"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
extend-exclude = '''
/(
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["migrations"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "heureum_platform.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "heureum_platform.settings"
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["."]
